Create door tunnel navigation hook
